+++ source/slang/slang.cpp
@@ -4858,10 +4858,10 @@ SlangResult EndToEndCompileRequest::getEntryPoint(SlangInt entryPointIndex, slan
     return SLANG_OK;
 }
 
-SlangResult EndToEndCompileRequest::isParameterLocationUsed(int entryPointIndex, int targetIndex, SlangParameterCategory category, int spaceIndex, int registerIndex, bool& outUsed)
+SlangResult EndToEndCompileRequest::isParameterLocationUsed(Int entryPointIndex, Int targetIndex, SlangParameterCategory category, UInt spaceIndex, UInt registerIndex, bool& outUsed)
 {
     CompileResult* compileResult = nullptr;
-    if (_getEntryPointResult(this, entryPointIndex, targetIndex, &compileResult) != SLANG_OK)
+    if (_getEntryPointResult(this, static_cast<int>(entryPointIndex), static_cast<int>(targetIndex), &compileResult) != SLANG_OK)
         return SLANG_E_INVALID_ARG;
 
     return compileResult->isParameterLocationUsed(category, spaceIndex, registerIndex, outUsed);
